split code in 3 files
